+++ lib/serega/validations/attribute/check_opt_delegate.rb
@@ -32,7 +32,7 @@ class Serega
 
             delegate_opts = opts[:delegate]
             check_opt_delegate_to(delegate_opts)
-            check_opt_delegate_key(delegate_opts)
+            check_opt_delegate_method(delegate_opts)
             check_opt_delegate_allow_nil(delegate_opts)
           end
 
@@ -43,8 +43,8 @@ class Serega
             Utils::CheckOptIsStringOrSymbol.call(delegate_opts, :to)
           end
 
-          def check_opt_delegate_key(delegate_opts)
-            Utils::CheckOptIsStringOrSymbol.call(delegate_opts, :key)
+          def check_opt_delegate_method(delegate_opts)
+            Utils::CheckOptIsStringOrSymbol.call(delegate_opts, :method)
           end
 
           def check_opt_delegate_allow_nil(delegate_opts)
@@ -52,7 +52,7 @@ class Serega
           end
 
           def check_usage_with_other_params(opts, block)
-            raise SeregaError, "Option :delegate can not be used together with option :key" if opts.key?(:key)
+            raise SeregaError, "Option :delegate can not be used together with option :method" if opts.key?(:method)
             raise SeregaError, "Option :delegate can not be used together with option :const" if opts.key?(:const)
             raise SeregaError, "Option :delegate can not be used together with option :value" if opts.key?(:value)
             raise SeregaError, "Option :delegate can not be used together with block" if block